LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/stretch/replaced-stretch-height-001-expected.txt


FAIL [data-expected-height], [data-expected-width] 1 assert_equals:
<canvas width="50" height="50" style="min-height: stretch; background: teal" data-expected-height="200" data-expected-width="200"></canvas>
width expected 200 but got 50
